pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "markets"
version = "0.2.0"
description = "Comprehensive market analysis toolkit with data retrieval, technical analysis, and fundamental analysis"
readme = "readme.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Michael Sands"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "requests>=2.26.0",
    "TA-Lib>=0.4.24",
    "yfinance>=0.2.0",
    "scipy>=1.7.0",
    "statsmodels>=0.13.0",
    "python-dateutil>=2.8.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "black>=21.0",
    "flake8>=3.9.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]
options = [
    "QuantLib>=1.28",
]
full = [
    "pytest>=6.2.0",
    "black>=21.0",
    "flake8>=3.9.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "QuantLib>=1.28",
]

[project.urls]
Homepage = "https://github.com/sandsmichael/submodules"
Repository = "https://github.com/sandsmichael/submodules"

[tool.setuptools]
packages = ["markets", "markets.submodules"]

[tool.setuptools.package-data]
markets = ["*.json"]
"markets.submodules" = ["*.json"]
